refactor(mobile): move the four inline test modules to sibling files (#1918)
Second crate of the per-crate migration campaign. Pure move via the
#[path] sibling pattern; test list verified byte-identical before/after
(124 lib tests). query.rs already had a #[path] sibling
(query_tests.rs, integration tests), so its inline unit tests land in
query_unit_tests.rs instead of colliding with it. Baseline shrinks by
the four velesdb-mobile lines (524 frozen lines paid).
